Tree toggle for header image now in header component. Needs styling and to change to non-toggle
