+++ src/components/Header/Header.js
@@ -5,15 +5,21 @@ import AddNewTodo from '../AddNewTodo/AddNewTodo';
 import TodoDate from '../Date/Date';
 import Progress from '../Progress/Progress';
 import HeaderImage from '../HeaderImage/HeaderImage';
+import tree from '../../images/pine.png';
 
 /**
  * Holder for the images which change according to progress,
  * the progress trackers and add new todo button/form
  */
-const Header = ({ items, addTodo, treeToggle }) => {
+const Header = ({ items, addTodo }) => {
   const [isFormDisplay, setIsFormDisplay] = useState(false);
+  const [treeToggle, setTreeToggle] = useState(false);
+
   const toggleIsFormDisplay = () => {
     setIsFormDisplay(!isFormDisplay);
+  };
+  const toggleTreeImage = () => {
+    setTreeToggle(!treeToggle);
   };
   let doneCount = 0;
 
@@ -37,6 +43,13 @@ const Header = ({ items, addTodo, treeToggle }) => {
         />
         <Progress donePercent={donePercent} />
         <TodoDate />
+        <input
+          type="image"
+          alt="tree header image toggle"
+          className={styles.treeButton}
+          src={tree}
+          onClick={toggleTreeImage}
+        />
       </div>
     </div>
   );
@@ -47,5 +60,4 @@ export default Header;
 Header.propTypes = {
   items: PropTypes.oneOfType([PropTypes.array]).isRequired,
   addTodo: PropTypes.func.isRequired,
-  treeToggle: PropTypes.bool.isRequired,
 };
