+++ src/components/Header/Header.module.css
@@ -10,3 +10,8 @@
   width: 480px;
   height: 270px;
 }
+.treeButton {
+  height: 20px;
+  margin-top: 10px;
+  margin-bottom: -15px;
+}
